Bingo framework guide
